requirements.txt
alabaster==1.0.0
asttokens==3.0.0
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.0
docutils==0.21.2
esbonio==0.16.5
executing==2.2.0
idna==3.10
imagesize==1.4.1
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jsonpickle==4.0.2
lsprotocol==2023.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
networkx==3.4.2
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
pure_eval==0.2.3
pygls==1.3.1
Pygments==2.18.0
pyspellchecker==0.8.2
pyvis==0.3.2
PyYAML==6.0.2
requests==2.32.3
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-codeautolink==0.15.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.3.0
wcwidth==0.2.13
